ForgeTrace (Repo-Analyser)/requirements.txt
# Core requirements
pyyaml>=6.0
jinja2>=3.1.0
gitpython>=3.1.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0

# ML dependencies for intelligent IP classification
# Install with: pip install -r requirements.txt
scikit-learn>=1.3.0  # Random Forest classifier, cross-validation
numpy>=1.24.0  # Numerical array operations
joblib>=1.3.0  # Safe model serialization compatible with sklearn
mlflow==3.6.0  # Tracking/training orchestration (pinned pending CVE fixes)

# Optional: Install with pip install -e ".[full]"
# py-tlsh>=4.7.2
# ppdeep
# weasyprint>=59.0
